LPD-51442 - Split tests
