+++ modules/test/playwright/tests/frontend-data-set-web/main/item_selection.spec.ts
@@ -1,0 +1,484 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {expect, mergeTests} from '@playwright/test';
+
+import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
+import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
+import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../fixtures/loginTest';
+import {fdsSamplePageTest} from './fixtures/fdsSamplePageTest';
+
+const test = mergeTests(
+	apiHelpersTest,
+	fdsSamplePageTest,
+	featureFlagsTest({
+		'LPD-41774': {enabled: true},
+		'LPS-178052': {enabled: true},
+	}),
+	isolatedSiteTest,
+	loginTest()
+);
+
+test.beforeEach(async ({fdsSamplePage, page, site}) => {
+	await fdsSamplePage.selectTab('Advanced');
+
+	await expect(
+		page.getByText('This is a description for sample 1.')
+	).toBeVisible();
+});
+
+test(
+	'Item selection and bulk actions',
+	{tag: ['@LPD-52063', '@LPD-41774']},
+	async ({fdsSamplePage, page}) => {
+		const firstItemRow = fdsSamplePage.table.bodyRows.first();
+
+		const firstItemCheckbox = firstItemRow
+			.locator('.cell-select-item')
+			.getByRole('checkbox');
+
+		await test.step('Select the first item in the table', async () => {
+			await firstItemCheckbox.check();
+
+			await expect(firstItemRow).toHaveClass('table-active');
+		});
+
+		await test.step('Check the highlighted bulk action "Label" is visible', async () => {
+			await expect(
+				fdsSamplePage.bulkActions.container.getByRole('button', {
+					name: 'Label',
+				})
+			).toHaveText('Label');
+		});
+
+		await test.step('Open ellipsis actions menu', async () => {
+			await fdsSamplePage.bulkActions.actionsDropdownButton.click();
+		});
+
+		await test.step('Check the bulk actions are listed', async () => {
+			await expect(
+				page.locator('.dropdown-menu.show').getByRole('menuitem')
+			).toHaveCount(3);
+			await expect(
+				page.locator('.dropdown-menu.show').getByRole('menuitem')
+			).toHaveText(['Label', 'Delete', 'Test']);
+		});
+
+		await test.step('Close ellipsis actions menu', async () => {
+			await fdsSamplePage.bulkActions.actionsDropdownButton.click();
+
+			await expect(page.locator('.dropdown-menu.show')).toBeHidden();
+		});
+
+		await test.step('Check in medium-width windows the bulk actions text is hidden', async () => {
+			await page.setViewportSize({height: 1024, width: 800});
+
+			const visibleLabelButton = fdsSamplePage.bulkActions.container
+				.locator('button')
+				.filter({
+					hasText: 'Label',
+				});
+
+			await expect(visibleLabelButton).toBeVisible();
+
+			await expect(
+				fdsSamplePage.bulkActions.container.getByLabel('Label')
+			).not.toBeVisible();
+		});
+
+		await test.step('Check in small-width windows the text and icon are hidden', async () => {
+			await page.setViewportSize({height: 720, width: 360});
+
+			await expect(
+				fdsSamplePage.bulkActions.container.getByRole('button', {
+					name: 'Label',
+				})
+			).toBeHidden();
+		});
+
+		await test.step('Reset the window size', async () => {
+			await page.setViewportSize({height: 720, width: 1280});
+		});
+
+		await test.step('Deselect the item to reset to original state', async () => {
+			await firstItemCheckbox.uncheck();
+		});
+
+		const itemsSelectorCheckbox = page.locator(
+			'input[name="items-selector"]'
+		);
+		let sentItems: Array<number>;
+		let sentKeyValues: Array<number>;
+		let sentSelectAll: boolean;
+
+		await page.route('/o/c/fdssamples/', async (route, request) => {
+			if (request.method() === 'POST') {
+				const postData = request.postDataJSON();
+
+				sentItems = postData.items;
+				sentKeyValues = postData.keyValues;
+				sentSelectAll = postData.selectAll;
+			}
+
+			await route.continue();
+		});
+
+		await test.step('Check Select All buton selects all elements', async () => {
+			await itemsSelectorCheckbox.click();
+
+			await expect(
+				page.getByText('20 of 75 Items Selected')
+			).toBeVisible();
+
+			await fdsSamplePage.selectAllCheckbox.click();
+
+			await expect(
+				page.getByText('All Selected (75 of 75 Items)')
+			).toBeVisible();
+		});
+
+		await test.step('Deselect an element disables Select All flag', async () => {
+			await fdsSamplePage.table.container
+				.locator('tbody .cell-select-item')
+				.first()
+				.getByRole('checkbox')
+				.uncheck();
+
+			await expect(
+				page.getByText('19 of 75 Items Selected')
+			).toBeVisible();
+
+			await expect(fdsSamplePage.selectAllCheckbox).not.toBeVisible();
+		});
+
+		await test.step('Without Select All flag active, requests sent actual item selection to bulk actions', async () => {
+			await fdsSamplePage.bulkActions.actionsDropdownButton.click();
+
+			await page
+				.locator('.dropdown-menu.show')
+				.getByRole('menuitem', {name: 'test'})
+				.click();
+
+			expect(sentItems).toHaveLength(19);
+			expect(sentKeyValues).toHaveLength(19);
+			expect(sentSelectAll).toBe(false);
+
+			expect(
+				await fdsSamplePage.bulkActions.actionsDropdownButton.getAttribute(
+					'aria-expanded'
+				)
+			).toBe('false');
+		});
+
+		await test.step('With Select All flag active, requests sent the flag instead of selected items', async () => {
+			await itemsSelectorCheckbox.click();
+
+			await fdsSamplePage.selectAllCheckbox.click();
+
+			await fdsSamplePage.bulkActions.actionsDropdownButton.click();
+
+			await page
+				.locator('.dropdown-menu.show')
+				.getByRole('menuitem', {name: 'test'})
+				.click();
+
+			expect(sentItems).toEqual([]);
+			expect(sentKeyValues).toEqual([]);
+			expect(sentSelectAll).toBe(true);
+
+			expect(
+				await fdsSamplePage.bulkActions.actionsDropdownButton.getAttribute(
+					'aria-expanded'
+				)
+			).toBe('false');
+
+			await fdsSamplePage.selectionToolbar.clearButton.click();
+		});
+
+		const firstListItem = fdsSamplePage.list.items.first();
+
+		const firstListItemCheckbox = firstListItem.getByRole('checkbox');
+
+		await test.step('Open list visualization mode', async () => {
+			await fdsSamplePage.changeVisualizationMode('List');
+		});
+
+		await test.step('Select the first item in the list', async () => {
+			await firstListItemCheckbox.check();
+
+			await expect(firstListItem).toHaveClass(/active/);
+		});
+	}
+);
+
+test(
+	'Check behavior of quick actions',
+	{tag: '@LPS-153220'},
+	async ({fdsSamplePage, page}) => {
+		const firstRowItemActionButton = fdsSamplePage.table.itemActionsCells
+			.first()
+			.getByRole('button', {
+				exact: true,
+				name: 'Actions',
+			});
+
+		const thirdRowItemActionButton = fdsSamplePage.table.itemActionsCells
+			.nth(2)
+			.getByRole('button', {
+				exact: true,
+				name: 'Actions',
+			});
+
+		const firstRowSampleEditQuickActionLink = fdsSamplePage.table.bodyRows
+			.first()
+			.getByLabel('Sample Edit');
+
+		const firstTableHeadCell = fdsSamplePage.table.headerCells.first();
+
+		await test.step('Assert that "#test-pencil" is appended to browser URL after clicking', async () => {
+			await firstTableHeadCell.hover();
+
+			await firstTableHeadCell.click();
+
+			await firstRowItemActionButton.hover();
+
+			await firstRowSampleEditQuickActionLink.click();
+
+			expect(page.url()).toContain('#test-pencil');
+		});
+
+		await test.step('Assert that clicking quick action is equivalent to clicking the ellipsis dropdown menu', async () => {
+			await firstRowItemActionButton.hover();
+
+			await firstRowSampleEditQuickActionLink.click();
+
+			const pageURLAfterQuickAction = page.url();
+
+			expect(pageURLAfterQuickAction).toContain('#test-pencil');
+
+			await firstRowItemActionButton.click();
+
+			await page
+				.getByRole('menuitem', {
+					name: 'Sample Edit',
+				})
+				.click();
+
+			expect(page.url()).toEqual(pageURLAfterQuickAction);
+		});
+
+		await test.step('Assert that hover over mouse off of the table body quick action menu is not visible', async () => {
+			await firstRowItemActionButton.hover();
+
+			await expect(firstRowSampleEditQuickActionLink).toBeVisible();
+
+			await firstTableHeadCell.hover();
+
+			await expect(firstRowSampleEditQuickActionLink).not.toBeVisible();
+		});
+
+		await test.step('When hovering over the first line item and the quick action menu is displayed on the 1st line', async () => {
+			const firstTableRow = fdsSamplePage.table.bodyRows.first();
+
+			await firstTableRow.hover();
+
+			await expect(firstRowSampleEditQuickActionLink).toBeVisible();
+		});
+
+		await test.step('When clicking on the ellipsis and hovering over another row, multiple quick action menus are displayed', async () => {
+			await thirdRowItemActionButton.click();
+
+			await expect(page.locator('.dropdown-menu.show')).toBeVisible();
+
+			await fdsSamplePage.table.bodyRows.first().hover();
+
+			await expect(firstRowSampleEditQuickActionLink).toBeVisible();
+
+			await firstTableHeadCell.click(); // Close dropdown
+		});
+
+		await test.step('Assert quick action can be displayed on only one active row', async () => {
+			await firstRowItemActionButton.hover();
+
+			await expect(firstRowSampleEditQuickActionLink).toBeVisible();
+
+			await thirdRowItemActionButton.hover();
+
+			await thirdRowItemActionButton.click();
+
+			await expect(firstRowSampleEditQuickActionLink).not.toBeVisible();
+
+			await firstTableHeadCell.click(); // Close dropdown
+		});
+
+		await test.step('Assert that quick action icons list should be limited to three actions', async () => {
+			await firstRowItemActionButton.hover();
+
+			await expect(
+				fdsSamplePage.table.bodyRows.first().getByLabel('View Details')
+			).toBeVisible();
+
+			await expect(
+				fdsSamplePage.table.bodyRows.first().getByLabel('Sample View')
+			).toBeVisible();
+
+			await expect(
+				fdsSamplePage.table.bodyRows.first().getByLabel('Sample Edit')
+			).toBeVisible();
+
+			await expect(
+				fdsSamplePage.table.bodyRows.first().getByLabel('Sample Copy')
+			).not.toBeVisible();
+		});
+
+		await test.step('Assert the quick action is not visible when the row checkbox is checked', async () => {
+			await fdsSamplePage.table.bodyRows
+				.first()
+				.getByRole('checkbox')
+				.click();
+
+			await firstRowItemActionButton.hover();
+
+			await expect(firstRowSampleEditQuickActionLink).not.toBeVisible();
+		});
+	}
+);
+
+test('InfoPanel behavior', async ({fdsSamplePage, page}) => {
+	const firstItemCheckbox = fdsSamplePage.table.container
+		.locator('tbody .cell-select-item')
+		.first()
+		.getByRole('checkbox');
+
+	const secondItemCheckbox = fdsSamplePage.table.container
+		.locator('tbody .cell-select-item')
+		.nth(1)
+		.getByRole('checkbox');
+
+	await test.step('Can open Info Panel when no item is selected', async () => {
+		expect(fdsSamplePage.toggleInfoPanelButton).toBeVisible();
+
+		await fdsSamplePage.toggleInfoPanelButton.click();
+
+		expect(fdsSamplePage.infoPanel).toBeInViewport();
+
+		expect(
+			page.getByText('Content from propsTransformer: No items selected')
+		).toBeVisible();
+
+		await fdsSamplePage.toggleInfoPanelButton.click();
+
+		expect(fdsSamplePage.infoPanel).not.toBeInViewport();
+	});
+
+	await test.step('Can open Info Panel when there is one item selected', async () => {
+		await firstItemCheckbox.check();
+
+		expect(fdsSamplePage.toggleInfoPanelButton).toBeVisible();
+
+		await fdsSamplePage.toggleInfoPanelButton.click();
+
+		expect(fdsSamplePage.infoPanel).toBeInViewport();
+
+		expect(
+			fdsSamplePage.infoPanel.getByText(
+				'This is a description for sample 1.'
+			)
+		).toBeVisible();
+
+		await fdsSamplePage.toggleInfoPanelButton.click();
+
+		expect(fdsSamplePage.infoPanel).not.toBeInViewport();
+	});
+
+	await test.step('Can open Info Panel when there are more than one items selected', async () => {
+		await firstItemCheckbox.check();
+
+		await secondItemCheckbox.check();
+
+		await fdsSamplePage.toggleInfoPanelButton.click();
+
+		expect(fdsSamplePage.infoPanel).toBeInViewport();
+
+		expect(
+			fdsSamplePage.infoPanel.getByText(
+				'Content from propsTransformer. Items selected: 2'
+			)
+		).toBeVisible();
+
+		await fdsSamplePage.toggleInfoPanelButton.click();
+
+		expect(fdsSamplePage.infoPanel).not.toBeInViewport();
+	});
+});
+
+test(
+	'Check selection style behavior',
+	{tag: '@LPD-49159'},
+	async ({fdsSamplePage, page}) => {
+		await test.step('Change visualization mode to List', async () => {
+			await fdsSamplePage.selectVisualizationMode('list');
+
+			const listItem = fdsSamplePage.list.container
+				.locator('.list-group-item')
+				.first();
+
+			await listItem.click();
+
+			await expect(listItem).toHaveClass(/active/);
+
+			await listItem.click();
+
+			await expect(listItem).not.toHaveClass(/active/);
+
+			await fdsSamplePage.clickItemAction('View Details');
+
+			await expect(listItem).toHaveClass(/active/);
+		});
+
+		await test.step('Change visualization mode to Cards', async () => {
+			await page.getByText('Clear').click();
+
+			await fdsSamplePage.selectVisualizationMode('cards');
+
+			const cardItem = fdsSamplePage.cards.container
+				.locator('.file-card')
+				.first();
+
+			await cardItem.click();
+
+			await expect(cardItem).toHaveClass(/active/);
+
+			await cardItem.click();
+
+			await expect(cardItem).not.toHaveClass(/active/);
+
+			await fdsSamplePage.clickItemAction('View Details');
+
+			await expect(cardItem).toHaveClass(/active/);
+		});
+
+		await test.step('Change visualization mode to Table', async () => {
+			await page.getByText('Clear').click();
+
+			await fdsSamplePage.selectVisualizationMode('customizedTable');
+
+			const tableItem = fdsSamplePage.table.bodyRows.nth(0);
+
+			await tableItem.click();
+
+			await expect(tableItem).toHaveClass(/active/);
+
+			await tableItem.click();
+
+			await expect(tableItem).not.toHaveClass(/active/);
+
+			await fdsSamplePage.clickItemAction('View Details');
+
+			await expect(tableItem).toHaveClass(/active/);
+		});
+	}
+);
